requirements.txt
streamlit>=1.28.0
jinja2>=3.1.2
pandas>=2.0.0
openpyxl>=3.1.0
weasyprint>=59.0
python-docx>=0.8.11
PyPDF2>=3.0.0
requests>=2.31.0